Start adding backend functionality to handle request priorities
My personal computer is having issues with Gradle... I'm getting an exception whenever I try to run or build the server (java.lang.IllegalArgumentException: Unsupported class file major version 63) in the settings file. I am really not sure what's up with that. I'll work on this more in the lab.